+++ tests/unit/test_fw_utils.py
@@ -14,7 +14,7 @@ import inspect
 from mock import Mock
 import os
 from sagemaker.fw_utils import create_image_uri, framework_name_from_image, framework_version_from_tag
-from sagemaker.fw_utils import tar_and_upload_dir, parse_s3_url, UploadedCode
+from sagemaker.fw_utils import tar_and_upload_dir, parse_s3_url, UploadedCode, validate_source_dir
 import pytest
 
 
@@ -90,34 +90,25 @@ def test_tar_and_upload_dir_s3(sagemaker_session):
     assert result == UploadedCode('s3://m', 'mnist.py')
 
 
-def test_tar_and_upload_dir_does_not_exits(sagemaker_session):
-    bucket = 'mybucker'
-    s3_key_prefix = 'something/source'
+def test_validate_source_dir_does_not_exits(sagemaker_session):
     script = 'mnist.py'
     directory = ' !@#$%^&*()path probably in not there.!@#$%^&*()'
-    with pytest.raises(ValueError) as error:
-        tar_and_upload_dir(sagemaker_session, bucket, s3_key_prefix, script, directory)
-    assert 'does not exist' in str(error)
+    with pytest.raises(ValueError):
+        validate_source_dir(script, directory)
 
 
-def test_tar_and_upload_dir_is_not_directory(sagemaker_session):
-    bucket = 'mybucker'
-    s3_key_prefix = 'something/source'
+def test_validate_source_dir_is_not_directory(sagemaker_session):
     script = 'mnist.py'
     directory = inspect.getfile(inspect.currentframe())
-    with pytest.raises(ValueError) as error:
-        tar_and_upload_dir(sagemaker_session, bucket, s3_key_prefix, script, directory)
-    assert 'is not a directory' in str(error)
+    with pytest.raises(ValueError):
+        validate_source_dir(script, directory)
 
 
-def test_tar_and_upload_dir_file_not_in_dir(sagemaker_session):
-    bucket = 'mybucker'
-    s3_key_prefix = 'something/source'
+def test_validate_source_dir_file_not_in_dir():
     script = ' !@#$%^&*() .myscript. !@#$%^&*() '
     directory = '.'
-    with pytest.raises(ValueError) as error:
-        tar_and_upload_dir(sagemaker_session, bucket, s3_key_prefix, script, directory)
-    assert 'No file named' in str(error)
+    with pytest.raises(ValueError):
+        validate_source_dir(script, directory)
 
 
 def test_tar_and_upload_dir_not_s3(sagemaker_session):