+++ CHANGELOG.rst
@@ -2,11 +2,13 @@
 CHANGELOG
 =========
 
+
 1.2.dev5
 ========
 
 * bug-fix: Change module names to string type in __all__
 * feature: Save training output files in local mode
+* feature: Local Mode: add support for local training data using file://
 
 1.2.4
 =====